backend/App/config.py
import os
from dotenv import load_dotenv

load_dotenv()
BASE_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))

class Config:
    SECRET_KEY=os.getenv('SECRET_KEY')
    SQLALCHEMY_DATABASE_URI = os.getenv('DATABASE_URL')
    SQLALCHEMY_TRACK_MODIFICATIONS = False
    UPLOAD_PROFILE_PICTURES = os.path.join(
        BASE_DIR,
        "uploads",
        "profile_pictures"
    )

SECRET_KEY = Config.SECRET_KEY