backend/requirements.txt
flask
python-dotenv
Flask_CORS
Flask_SQLAlchemy
Flask_Migrate
psycopg2-binary
PyJWT
geoalchemy2
geoalchemy2[shapely]
validate-docbr